Move security check code from FCMHandlerService to SecurityPolicyChecker
